+++ src/main/java/com/poortorich/iteration/validator/EndValidator.java
@@ -4,6 +4,7 @@ import com.poortorich.iteration.constants.IterationResponseMessages;
 import com.poortorich.iteration.constants.IterationValidationConstraints;
 import com.poortorich.iteration.entity.enums.EndType;
 import com.poortorich.iteration.request.End;
+import com.poortorich.iteration.validator.annotations.EndCheck;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
