+++ src/main/java/com/poortorich/iteration/validator/IterationRuleValidator.java
@@ -5,6 +5,7 @@ import com.poortorich.iteration.constants.IterationValidationConstraints;
 import com.poortorich.iteration.entity.enums.IterationRuleType;
 import com.poortorich.iteration.request.IterationRule;
 import com.poortorich.iteration.request.MonthlyOption;
+import com.poortorich.iteration.validator.annotations.IterationRuleCheck;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
