+++ src/main/java/com/poortorich/iteration/validator/annotations/MonthlyOptionCheck.java
@@ -1,6 +1,7 @@
-package com.poortorich.iteration.validator;
+package com.poortorich.iteration.validator.annotations;
 
 import com.poortorich.iteration.constants.IterationResponseMessages;
+import com.poortorich.iteration.validator.MonthlyOptionValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -11,8 +12,8 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = IterationRuleValidator.class)
-public @interface IterationRuleCheck {
+@Constraint(validatedBy = MonthlyOptionValidator.class)
+public @interface MonthlyOptionCheck {
 
     String message() default IterationResponseMessages.CHECK_DEFAULT_MESSAGE;
     Class<?>[] groups() default {};
